fix: 스터디 게시판 페이지 api 함수로 분리
